+++ src/Endpoint/HelloMediatR/Controllers/AuthActions/AuthController_RefreshToken.cs
@@ -21,35 +21,6 @@ namespace Hello.MediatR.Endpoint.Controllers.AuthActions
         [Route("api/auth/refresh-token")] // latest
         [HttpPost]
         [SwaggerResponse((int)HttpStatusCode.OK, Type = typeof(ISvcResponseBaseDto<JwtTokenDto>))]
-        public async Task<IActionResult> RefreshToken_latest([FromBody] AuthRefreshToken_RequestDto requestDto) => await RefreshToken_v202104(requestDto);
-
-        /// <summary>
-        /// Refresh token key
-        /// </summary>
-        /// <param name="requestDto"></param>
-        /// <returns></returns>
-        [ApiVersion("202104")]
-        [ApiExplorerSettings(GroupName = "v202104")]
-        [Route("api/v202104/auth/refresh-token")]
-        [HttpPost]
-        [SwaggerResponse((int)HttpStatusCode.OK, Type = typeof(ISvcResponseBaseDto<JwtTokenDto>))]
-        public async Task<IActionResult> RefreshToken_v202104([FromBody] AuthRefreshToken_RequestDto requestDto) => await RefreshToken_v202103(requestDto);
-
-        /// <summary>
-        /// Refresh token key
-        /// </summary>
-        /// <param name="requestDto"></param>
-        /// <returns></returns>
-        [ApiVersion("202103")]
-        [ApiExplorerSettings(GroupName = "v202103")]
-        [Route("api/v202103/auth/refresh-token")]
-        [HttpPost]
-        [SwaggerResponse((int)HttpStatusCode.OK, Type = typeof(ISvcResponseBaseDto<JwtTokenDto>))]
-        public async Task<IActionResult> RefreshToken_v202103([FromBody] AuthRefreshToken_RequestDto requestDto)
-        {
-            var res = _sessionAdmin.RefeshToken(requestDto.RefreshToken);
-            await Task.CompletedTask;
-            return Ok(res);
-        }
+        public async Task<IActionResult> RefreshToken_latest([FromBody] AuthRefreshToken_RequestDto requestDto) => await RefreshToken_v202103(requestDto);
     }
 }