+++ ChangeLog.md
@@ -1,3 +1,6 @@
 ### 2021-04-18
 * **Hello.MediatR.Endpoint (1.0.0)**
   * Created
+### 2023-03-28
+* **Hello.MediatR.Endpoint (2.0.0)**
+  * Upgrade to NET 7.0 from NET 5.0